backend/requirements.txt
amqp==2.3.2
billiard==3.5.0.3
celery==4.2.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-cors-headers==2.3.0
django-redis-cache==1.7.1
djangorestframework==3.8.2
djangorestframework-simplejwt==3.2.3
idna==2.7
kombu==4.2.1
PyJWT==1.6.4
pytz==2018.4
redis==2.10.6
requests==2.19.1
urllib3==1.23
vine==1.1.4
